dcache-info: clean code, improve readability and logged values
Motivation:
Modification:
Result:

Cleaner code, improved readability, improved accuracy of logged time until next trigger value.

Target: master
Requires-notes: no
Requires-book: no
Patch: https://rb.dcache.org/r/13568/
Acked-by: Albert Rossi

+++ modules/dcache-info/src/main/java/org/dcache/services/info/gathers/srm/LinkgroupDetailsDga.java
@@ -41,7 +41,7 @@ public class LinkgroupDetailsDga extends SkelPeriodicActivity {
      *
      * @param interval how often the list of linkgroups should be updated, in seconds.
      */
-    public LinkgroupDetailsDga(CellPath spacemanager, MessageSender sender, int interval) {
+    public LinkgroupDetailsDga(CellPath spacemanager, MessageSender sender, long interval) {
         super(interval);
         _sender = sender;
         _metricLifetime = Math.round(interval * SAFETY_FACTOR);
